Refactoring TripItems code to Buttons and Toggles, added GeoControl ref for auto ask to locate, working on New Trip
